correção de tamanho tabela

+++ pages/2_LANÇAMENTOS.py
@@ -98,7 +98,7 @@ if tamanho_tabela==2:
 
 else:
   lançamentos['BANCO'] = lançamentos['BANCO'].str.upper()
-  #lançamentos['VALOR'] = lançamentos['VALOR'].astype(float)
+  
   lançamentos = lançamentos.set_index('ID')
   colunas = list(lançamentos.columns)
   colunas_selecionadas = st.multiselect('Selecione as colunas da tabela:', colunas, colunas,)
@@ -118,7 +118,7 @@ else:
   pesqdescri = st.sidebar.text_input('Pesquisar descrição')
 
 
-
+  lançamentos['VALOR'] = lançamentos['VALOR'].astype(float)
   lançamentos = lançamentos[lançamentos['BANCO'].notna()]
   lançamentos['DATA'] = pd.to_datetime(lançamentos['DATA'], dayfirst=True, errors='coerce')
   lançamentos_conciliados = lançamentos[lançamentos['CONCILIADO']==True]
